chore(utils): move utils directory
