+++ planning/behavior_path_planner/include/behavior_path_planner/path_utilities.hpp
@@ -15,8 +15,9 @@
 #ifndef BEHAVIOR_PATH_PLANNER__PATH_UTILITIES_HPP_
 #define BEHAVIOR_PATH_PLANNER__PATH_UTILITIES_HPP_
 
+#include "behavior_path_planner/util/path_shifter/path_shifter.hpp"
+
 #include <behavior_path_planner/parameters.hpp>
-#include <behavior_path_planner/scene_module/utils/path_shifter.hpp>
 
 #include <autoware_auto_planning_msgs/msg/path.hpp>
 #include <autoware_auto_planning_msgs/msg/path_with_lane_id.hpp>
