+++ planning/behavior_path_planner/include/behavior_path_planner/scene_module/pull_out/geometric_pull_out.hpp
@@ -17,7 +17,7 @@
 
 #include "behavior_path_planner/scene_module/pull_out/pull_out_path.hpp"
 #include "behavior_path_planner/scene_module/pull_out/pull_out_planner_base.hpp"
-#include "behavior_path_planner/scene_module/utils/geometric_parallel_parking.hpp"
+#include "behavior_path_planner/util/geometric_parallel_parking/geometric_parallel_parking.hpp"
 
 #include <autoware_auto_planning_msgs/msg/path_with_lane_id.hpp>
