+++ planning/behavior_path_planner/include/behavior_path_planner/util/path_shifter/path_shifter.hpp
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef BEHAVIOR_PATH_PLANNER__SCENE_MODULE__UTILS__PATH_SHIFTER_HPP_
-#define BEHAVIOR_PATH_PLANNER__SCENE_MODULE__UTILS__PATH_SHIFTER_HPP_
+#ifndef BEHAVIOR_PATH_PLANNER__UTIL__PATH_SHIFTER__PATH_SHIFTER_HPP_
+#define BEHAVIOR_PATH_PLANNER__UTIL__PATH_SHIFTER__PATH_SHIFTER_HPP_
 
 #include "behavior_path_planner/parameters.hpp"
 
@@ -226,4 +226,4 @@ private:
 
 }  // namespace behavior_path_planner
 
-#endif  // BEHAVIOR_PATH_PLANNER__SCENE_MODULE__UTILS__PATH_SHIFTER_HPP_
+#endif  // BEHAVIOR_PATH_PLANNER__UTIL__PATH_SHIFTER__PATH_SHIFTER_HPP_
