+++ planning/behavior_path_planner/src/scene_module/lane_change/debug.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "behavior_path_planner/scene_module/utils/path_shifter.hpp"
+#include "behavior_path_planner/util/path_shifter/path_shifter.hpp"
 
 #include <behavior_path_planner/scene_module/lane_change/debug.hpp>
 #include <tier4_autoware_utils/ros/marker_helper.hpp>
