+++ planning/behavior_path_planner/src/scene_module/pull_out/util.cpp
@@ -15,8 +15,8 @@
 #include "behavior_path_planner/scene_module/pull_out/util.hpp"
 
 #include "behavior_path_planner/path_utilities.hpp"
-#include "behavior_path_planner/scene_module/utils/path_shifter.hpp"
 #include "behavior_path_planner/util/create_vehicle_footprint.hpp"
+#include "behavior_path_planner/util/path_shifter/path_shifter.hpp"
 
 #include <lanelet2_extension/utility/utilities.hpp>
 #include <rclcpp/rclcpp.hpp>
