+++ planning/behavior_path_planner/src/scene_module/pull_over/util.cpp
@@ -15,7 +15,7 @@
 #include "behavior_path_planner/scene_module/pull_over/util.hpp"
 
 #include "behavior_path_planner/path_utilities.hpp"
-#include "behavior_path_planner/scene_module/utils/path_shifter.hpp"
+#include "behavior_path_planner/util/path_shifter/path_shifter.hpp"
 
 #include <lanelet2_extension/utility/query.hpp>
 #include <lanelet2_extension/utility/utilities.hpp>
